Reduced size of drawable to temporarily fix #5
Not the best solution though. Drawable size should be calculated through code based on multiple factors such as device screen size and ```withDistance``` value. But for now and until you have some time this should allow users to use the library (not without a HUGE abuse of memory due to it's implementation).